Don't persist MAKEFLAGS in the image
While working on an image deep in the inheritance hierarchy that's rooted at this image, it became clear to me that persisting MAKEFLAGS in the image (by using ENV) is not a great choice, because building some packages in parallel is problematic. We're better off letting each image define it as needed.