Avoid polluting Docker environment.
Don't pollute Docker environment with variables that are only needed for
building the image.

Also move ccache configuration out of the environment and into a config
file so that descendant images can revert to defaults more easily. The
ccache-related environment variables override all settings, and should
only be used as a last resort. This, together with the facts that (1)
Dockerfiles don't currently provide a mechanism for unsetting
environment variables, and (2) ccache complains when its environment
variables are set but empty, make it nigh impossible for descendant
images to reliably reset ccache to defaults when ccache's configuration
is baked into the image's environment.